renaming these to HotGlue::XYXTemplate
