+++ lib/generators/hot_glue/markup_templates/base.rb
@@ -1,0 +1,10 @@
+module  HotGlue
+  class TemplateBase
+    def initialize(build)
+      # pass the scaffold build object into this
+
+      @build_object = build
+    end
+
+  end
+end
