+++ lib/generators/hot_glue/markup_templates/erb.rb
@@ -1,0 +1,6 @@
+module  HotGlue
+  class ErbTemplate < TemplateBase
+
+  end
+
+end
