+++ lib/generators/hot_glue/markup_templates/slim.rb
@@ -1,0 +1,9 @@
+
+
+module  HotGlue
+  class Slim < TemplateBase
+
+
+
+  end
+end
