+++ lib/generators/hot_glue/scaffold_generator.rb
@@ -1,6 +1,10 @@
 require 'rails/generators/erb/scaffold/scaffold_generator'
 require 'ffaker'
 
+require_relative './markup_templates/base'
+require_relative './markup_templates/erb'
+require_relative './markup_templates/haml'
+require_relative './markup_templates/slim'
 
 module HotGlue
   class Error < StandardError
@@ -42,13 +46,7 @@ module HotGlue
 
 
 
-    def field_output(col, type = nil, width, col_identifier )
 
-      "#{col_identifier}{class: \"form-group \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-    = f.text_field :#{col.to_s}, value: @#{singular}.#{col.to_s}, size: #{width}, class: 'form-control', type: '#{type}'
-    %label.form-text
-      #{col.to_s.humanize}\n"
-    end
   end
 
 
@@ -80,7 +78,7 @@ module HotGlue
     class_option :no_paginate, type: :boolean, default: false
     class_option :big_edit, type: :boolean, default: false
     class_option :show_only, type: :string, default: ""
-    # class_option :markup, type: :string, default: "erb"
+    class_option :markup, type: :string, default: "erb"
 
     def initialize(*meta_args) #:nodoc:
       super
@@ -96,18 +94,20 @@ module HotGlue
         raise(HotGlue::Error, "*** Oops: You seem to have specified both the --specs-only flag and --no-specs flags. this doesn't make any sense, so I am aborting. sorry.")
       end
 
-      # if options['markup'] == "erb"
-      #   puts "ERB IS NOT IMPLEMENTED"
-      #   abort
-      #   # raise "erb not implemented"
-      # elsif options['markup'] == "slim"
-      #   puts "SLIM IS NOT IMPLEMENTED"
-      #   abort
-      #
-      # elsif options['markup'] == "haml"
-      #   byebug
-      #   @template_builder = HotGlue::TemplateBuilders::Haml.new(self)
-      # end
+      if options['markup'] == "erb"
+        puts "ERB IS NOT IMPLEMENTED"
+        abort
+        # raise "erb not implemented"
+
+        @template_builder = ::Erb.new(self)
+      elsif options['markup'] == "slim"
+        puts "SLIM IS NOT IMPLEMENTED"
+        abort
+        @template_builder = ::Slim.new(self)
+
+      elsif options['markup'] == "haml"
+        @template_builder = HotGlue::HamlTemplate.new(self)
+      end
 
 
       args = meta_args[0]
@@ -562,93 +562,13 @@ module HotGlue
     end
 
     def all_form_fields
-      col_identifier = "  .col"
-      col_spaces_prepend = "    "
-
-      res = @columns.map { |col|
-
-        if @show_only.include?(col)
-
-          "#{col_identifier}{class: \"form-group \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-    = @#{singular}.#{col.to_s}
-    %label.form-text
-      #{col.to_s.humanize}\n"
-        else
-
-
-        type = eval("#{singular_class}.columns_hash['#{col}']").type
-        limit = eval("#{singular_class}.columns_hash['#{col}']").limit
-        sql_type = eval("#{singular_class}.columns_hash['#{col}']").sql_type
-
-        case type
-        when :integer
-          # look for a belongs_to on this object
-          if col.to_s.ends_with?("_id")
-            assoc_name = col.to_s.gsub("_id","")
-            assoc = eval("#{singular_class}.reflect_on_association(:#{assoc_name})")
-            if assoc.nil?
-              exit_message= "*** Oops. on the #{singular_class} object, there doesn't seem to be an association called '#{assoc_name}'"
-              exit
-            end
-            display_column = derrive_reference_name(assoc.class_name)
-
-
-            "#{col_identifier}{class: \"form-group \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{assoc_name.to_s})}\"}
-#{col_spaces_prepend}= f.collection_select(:#{col.to_s}, #{assoc.class_name}.all, :id, :#{display_column}, {prompt: true, selected: @#{singular}.#{col.to_s} }, class: 'form-control')
-#{col_spaces_prepend}%label.small.form-text.text-muted
-#{col_spaces_prepend}  #{col.to_s.humanize}"
-
-          else
-            "#{col_identifier}{class: \"form-group \#{'alert-danger' if @#{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-#{col_spaces_prepend}= f.text_field :#{col.to_s}, value: #{singular}.#{col.to_s}, class: 'form-control', size: 4, type: 'number'
-#{col_spaces_prepend}%label.form-text
-#{col_spaces_prepend}  #{col.to_s.humanize}\n"
-          end
-        when :string
-          limit ||= 256
-          if limit <= 256
-            field_output(col, nil, limit, col_identifier)
-          else
-            text_area_output(col, limit, col_identifier)
-          end
-
-        when :text
-          limit ||= 256
-          if limit <= 256
-            field_output(col, nil, limit, col_identifier)
-          else
-            text_area_output(col, limit, col_identifier)
-          end
-        when :float
-          limit ||= 256
-          field_output(col, nil, limit, col_identifier)
-
-
-        when :datetime
-          "#{col_identifier}{class: \"form-group \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-#{col_spaces_prepend}= datetime_field_localized(f, :#{col.to_s}, #{singular}.#{col.to_s}, '#{col.to_s.humanize}', #{@auth ? @auth+'.timezone' : 'nil'})"
-        when :date
-          "#{col_identifier}{class: \"form-group \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-#{col_spaces_prepend}= date_field_localized(f, :#{col.to_s}, #{singular}.#{col.to_s}, '#{col.to_s.humanize}', #{@auth ? @auth+'.timezone' : 'nil'})"
-        when :time
-          "#{col_identifier}{class: \"form-group  \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-#{col_spaces_prepend}= time_field_localized(f, :#{col.to_s}, #{singular}.#{col.to_s}, '#{col.to_s.humanize}', #{@auth ? @auth+'.timezone' : 'nil'})"
-        when :boolean
-          "#{col_identifier}{class: \"form-group  \#{'alert-danger' if #{singular}.errors.details.keys.include?(:#{col.to_s})}\"}
-#{col_spaces_prepend}%span
-#{col_spaces_prepend}  #{col.to_s.humanize}
-#{col_spaces_prepend}= f.radio_button(:#{col.to_s},  '0', checked: #{singular}.#{col.to_s}  ? '' : 'checked')
-#{col_spaces_prepend}= f.label(:#{col.to_s}, value: 'No', for: '#{singular}_#{col.to_s}_0')
-
-#{col_spaces_prepend}= f.radio_button(:#{col.to_s}, '1',  checked: #{singular}.#{col.to_s}  ? 'checked' : '')
-#{col_spaces_prepend}= f.label(:#{col.to_s}, value: 'Yes', for: '#{singular}_#{col.to_s}_1')
-      "
-        end
-        end
-      }.join("\n")
-      return res
+      @template_builder.all_form_fields(
+        columns: @columns,
+        show_only: @show_only,
+        singular_class: singular_class,
+        singular: singular
+      )
     end
-
 
     def all_line_fields
       columns = @columns.count + 1
