--- lib/generators/hot_glue/template_builders/base.rb
@@ -1,11 +1,0 @@
-module  HotGlue
-  module TemplateBuilders
-    class Base
-      def initialize(build)
-        # bass the build object into the template builder
-        @build = build
-      end
-
-    end
-  end
-end
