--- lib/generators/hot_glue/template_builders/erb.rb
@@ -1,6 +1,0 @@
-class HotGlue::TemplateBuilders::Erb < HotGlue::TemplateBuilders::Base
-
-
-
-
-end
