--- lib/generators/hot_glue/template_builders/slim.rb
@@ -1,9 +1,0 @@
-
-
-module HotGlue
-
-
-
-  class TemplateBuilders::Slim <  HotGlue::TemplateBuilders::Base
-  end
-end
