+++ lib/hot-glue.rb
@@ -3,13 +3,12 @@ require "hotglue/engine"
 require 'kaminari'
 require 'haml-rails'
 
-
 module HotGlue
 
 
+
+  module TemplateBuilders
+
+  end
 end
 
-require 'lib/generators/hot_glue/template_builders/base'
-require 'lib/generators/hot_glue/template_builders/erb'
-require 'lib/generators/hot_glue/template_builders/haml'
-require 'lib/generators/hot_glue/template_builders/slim'
